\name{MapTyrol}
\alias{MapTyrol}

\title{Topographic data for Tyrol}

\description{Topographic data to plot a map of Tyrol and surrounding areas.}

\details{Downloaded from \url{http://www.earthenv.org/DEM}}

\usage{data("MapTyrol", package = "RainTyrol")}

\format{
A \code{RasterLayer} containing topographic data of Tyrol and surrounding areas.
}

\references{
Robinson N, Regetz J, Guralnick R P (2014).
{{E}arth{E}nv-{DEM90}: A Nearly-Global, Void-Free, Multi-Scale Smoothed, 90m Digital Elevation Model From Fused {ASTER} and {SRTM} Data},
\emph{ISPRS Journal of Photogrammetry and Remote Sensing}, \bold{87}, 57--67.
    \doi{10.1016/j.isprsjprs.2013.11.002}


EarthEnv-DEM90e website: \url{http://www.earthenv.org/DEM.html}
}

\examples{
data("MapTyrol", package = "RainTyrol")
}

\keyword{datasets}
