\name{MapTyrol_border}
\alias{MapTyrol_border}
\encoding{UTF-8}

\title{Border of Tyrol}

\description{Data to plot the borders of Tyrol.}

\source{\url{https://www.data.gv.at/katalog/dataset/vgd-stichtagsdaten-1-250-000}}

\usage{data("MapTyrol_border")}

\format{
An object of class \code{SpatialPolygons} containing the border of Tyrol.
}

\references{
{B}undesamt für {E}ich- und {V}ermessungswesen

\url{https://www.data.gv.at/katalog/dataset/vgd-stichtagsdaten-1-250-000}
}

\examples{
data("MapTyrol_border", package = "RainTyrol")
}

\keyword{datasets}
